perf: remove optional chain
