+++ CHANGELOG.zh_CN.md
@@ -1,9 +1,21 @@
 ## Wip
 
+### ✨ Features
+
+- 面包屑支持显示图标
+- 新增 tinymce 富文本组件
+
+### 🎫 Chores
+
+- 删除代码内的可选链语法
+
 ### 🐛 Bug Fixes
 
 - 修复抽屉组件自动高度及显示 footer 显示问题
 - 修复表单查询后重置回默认值
+- 修复菜单没有子节点时显示折叠的问题
+- 修复面包屑显示样式问题
+- 修复 modal 在 destroyOnClose=true 时多次打开拖拽失效
 
 # 2.0.0-rc.4 (2020-10-21)
