+++ src/components/Authority/index.ts
@@ -1,0 +1,2 @@
+import Authority from './src/index.vue';
+export default Authority;
