+++ src/components/Basic/index.ts
@@ -1,3 +1,3 @@
 export { default as BasicArrow } from './src/BasicArrow.vue';
-export { default as BasicHelp } from './src/BasicHelp';
+export { default as BasicHelp } from './src/BasicHelp.vue';
 export { default as BasicTitle } from './src/BasicTitle.vue';
