--- src/components/Basic/src/BasicHelp.less
@@ -1,19 +1,0 @@
-@import (reference) '../../../design/index.less';
-
-.base-help {
-  display: inline-block;
-  margin-left: 6px;
-  font-size: 14px;
-  color: @text-color-help-dark;
-  cursor: pointer;
-
-  &:hover {
-    color: @primary-color;
-  }
-
-  &__wrap {
-    p {
-      margin-bottom: 0;
-    }
-  }
-}
